+++ CHANGELOG.md
@@ -8,6 +8,11 @@ title: Changelog
 
 -   `@includeCode` can now inject parts of files using region names or line
     numbers, #2816.
+-   The `@inline` tag now works in more places for generic types.
+
+### Bug Fixes
+
+-   Fixed an issue where TypeDoc would incorrectly ignore type arguments in references, #2823.
 
 ## v0.27.6 (2024-12-26)